Project Update: Basic Layout
The folder framework has now been setup and the scripts are now starting to be put in place. Victor should handle the layout of the screen and the camera position later. Just going to use 4 characters for the time being to test (2 friendly and 2 enemies). Should be able to scale the system up later.

From here now that I have some models to work with, I plan on transfering the code I have written down into the set scripts and get the basic battle system working.